Add metadata for all pages and separate client shells
